Extract 'retryAsync' to 'lib/util' directory
